# Add thread-local panic context for improved debugging