--- a/components/tikv_util/src/lib.rs
+++ b/components/tikv_util/src/lib.rs
@@ -11,10 +11,12 @@
 extern crate test;
 
 use std::{
+    cell::RefCell,
     cmp,
     collections::{
         hash_map::Entry,
         vec_deque::{Iter, VecDeque},
+        HashMap,
     },
     convert::AsRef,
     env,
@@ -441,6 +443,68 @@ impl<T> Drop for MustConsumeVec<T> {
     }
 }
 
+// Thread-local variable for storing the panic context.
+thread_local! {
+    pub static PANIC_CONTEXT: RefCell<HashMap<String, String>> = RefCell::new(HashMap::new());
+}
+
+/// Guard to automatically remove a key from `PANIC_CONTEXT` on drop.
+pub struct PanicContextGuard(pub Vec<String>);
+
+impl Drop for PanicContextGuard {
+    fn drop(&mut self) {
+        let _ = crate::PANIC_CONTEXT.try_with(|ctx| {
+            if let Ok(mut ctx) = ctx.try_borrow_mut() {
+                for key in self.0.iter() {
+                    ctx.remove(key);
+                }
+            }
+        });
+    }
+}
+
+/// Sets key-value pairs in the thread-local panic context that will be logged
+/// on panic. Each key is prefixed with the source code location.
+///
+/// Returns a `PanicContextGuard` to ensure the keys are removed on scope exit.
+#[macro_export]
+macro_rules! set_panic_context {
+    ( $($key:expr => $value:expr),+ $(,)?) => {{
+        let location = format!("{}:{}", file!(), line!());
+        let mut keys = Vec::new();
+        $(
+            let full_key = format!("({}) {}", location, $key);
+            let _ = $crate::PANIC_CONTEXT.try_with(|ctx| {
+                if let Ok(mut ctx) = ctx.try_borrow_mut() {
+                    ctx.insert(full_key.clone(), $value.to_string());
+                }
+            });
+            keys.push(full_key);
+        )+
+        $crate::PanicContextGuard(keys)
+    }};
+}
+
+/// Retrieves the current panic context as a string of key-value pairs.
+pub fn get_panic_context() -> String {
+    PANIC_CONTEXT
+        .try_with(|ctx| {
+            ctx.try_borrow()
+                .map(|ctx| {
+                    if ctx.is_empty() {
+                        String::new()
+                    } else {
+                        ctx.iter()
+                            .map(|(k, v)| format!("{}=>{}", k, v))
+                            .collect::<Vec<_>>()
+                            .join(", ")
+                    }
+                })
+                .unwrap_or_default()
+        })
+        .unwrap_or_default()
+}
+
 /// Exit the whole process when panic.
 pub fn set_panic_hook(panic_abort: bool, data_dir: &str) {
     use std::{panic, process};
@@ -481,6 +545,7 @@ pub fn set_panic_hook(panic_abort: bool, data_dir: &str) {
             "thread_name" => name,
             "location" => loc.unwrap_or_else(|| "<unknown>".to_owned()),
             "backtrace" => format_args!("{:?}", bt),
+            "context" => get_panic_context(),
         );
 
         // There might be remaining logs in the async logger.
@@ -808,5 +873,23 @@ mod tests {
             // the test would fail.
         });
         res.unwrap_err();
+    }
+
+    #[test]
+    fn test_panic_context() {
+        let _guard = set_panic_context! {"outer_ctx" => "foo"};
+        {
+            let _guard = set_panic_context! {
+                "inner_ctx1" => "bar",
+                "inner_ctx2" => "baz",
+            };
+            assert!(get_panic_context().contains("outer_ctx=>foo"));
+            assert!(get_panic_context().contains("inner_ctx1=>bar"));
+            assert!(get_panic_context().contains("inner_ctx2=>baz"));
+        } // Inner guard is dropped, removing "inner_ctx"
+
+        // Only the outer context remains
+        assert!(get_panic_context().contains("outer_ctx=>foo"));
+        assert!(!get_panic_context().contains("inner_ctx"));
     }
 }