<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed
-->

### What is changed and how it works?
<!--
Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#linking-issues.
-->

Issue Number: close #18039

<!--
You could use "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#format-of-the-commit-message.
-->
What's Changed:

```commit-message
This commit adds a thread-local panic context to log key debugging
details (e.g., `region_id`, `peer_id`) during panics. The
`set_panic_context` macro enables adding key-value pairs to the
panic context, with automatic cleanup when the returned guard goes
out of scope. The panic context is logged in the panic hook.

Currently, the panic context is set in the `handle_normal` function,
covering per-region raft processing. More context can be added as
needed in the future.
```

Credit to @glorv for suggesting the idea behind this approach!

### Related changes

- [ ] PR to update `pingcap/docs`/`pingcap/docs-cn`:
- [ ] Need to cherry-pick to the release branch

### Check List

Tests <!-- At least one of them must be included. -->
- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code


Logs from manual tests in which I intentionally injected panics into TiKV:
```
[2024/12/31 15:15:58.552 +08:00] [FATAL] [lib.rs:544] ["intentional panic"] [context="(src/server/snap.rs:381) snap_key=>40_6_27"] [backtrace="... <truncated> ..."] [location=/Users/bishengh/code/tikv/src/server/snap.rs:412] [thread_name=snap-sender] [thread_id=140]
```

```
[2024/12/31 15:23:35.162 +08:00] [FATAL] [lib.rs:544] ["intentional panic"] [context="(components/raftstore/src/store/fsm/store.rs:1077) peer_id=>5, (components/raftstore/src/store/fsm/store.rs:1077) region_id=>4"] [backtrace="... <truncated> ..."] [location=components/raftstore/src/store/entry_storage.rs:185] [thread_name=raftstore-1-1] [thread_id=101]
```


Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

### Release note
<!-- 
Compatibility change, improvement, bugfix, and new feature need a release note.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

If you don't think this PR needs a release note then fill it with None.
If this PR will be picked to release branch, then a release note is probably required.
-->

```release-note
None
```