@LykxSassinator 2024-12-31
Is there any risk of performance regression ? 

Because this logging of panic_context has added an additional pre-setting operation during the processing of `handle_normal`.