@hbisheng 2025-01-07
> Is there any risk of performance regression ?
> 
> Because this logging of panic_context has added an additional pre-setting operation during the processing of `handle_normal`.

I assumed memory operations are cheap but we should test it. Will attach some benchmark results.