@hbisheng 2025-02-06
cc @LykxSassinator 
Benchmark results showed minor fluctuations of QPS but no clear regression:

### OLTP read_write (Run 1)
| Test Case          | Duration | TPS  | QPS   | Avg Latency | p99 Latency | p999 Latency | QPS Change |
|--------------------|----------|------|-------|-------------|-------------|--------------|-------------|
| **Baseline**  | 30m      | 2525 | 49886 | 1.80ms      | 12.2ms      | 25.8ms       | -           |
| **PR**        | 30m      | 2497 | 49306 | 1.82ms      | 11.1ms      | 24.4ms       | **-1.2%**       |

### OLTP read_write (Run 2)
| Test Case          | Duration | TPS  | QPS   | Avg Latency | p99 Latency | p999 Latency | QPS Change |
|--------------------|----------|------|-------|-------------|-------------|--------------|-------------|
| **Baseline**  | 30m      | 2481 | 48986 | 1.81ms      | 11.9ms      | 27.8ms       | -           |
| **PR**        | 30m      | 2494 | 49257 | 1.81ms      | 11.8ms      | 27.0ms       | **+0.6%**       |

### TPCC (Run 1)
| Test Case          | Duration | TPS  | QPS   | Avg Latency | p99 Latency | p999 Latency | QPS Change |
|--------------------|----------|------|-------|-------------|-------------|--------------|-------------|
| **Baseline**  | 30m      | 2447 | 37707 | 2.51ms      | 30.4ms      | 61.5ms       | -           |
| **PR**        | 30m      | 2368 | 36476 | 2.61ms      | 31.6ms      | 66.4ms       | **-3.3%**       |

### TPCC (Run 2)
| Test Case          | Duration | TPS  | QPS   | Avg Latency | p99 Latency | p999 Latency | QPS Change |
|--------------------|----------|------|-------|-------------|-------------|--------------|-------------|
| **Baseline**  | 30m      | 2367 | 36456 | 2.58ms      | 32.4ms      | 65.5ms       | -           |
| **PR**        | 30m      | 2403 | 37023 | 2.54ms      | 31.5ms      | 62.7ms       | **+1.6%**       |